+++ src/cli/commands/connection-runtime.ts
@@ -21,7 +21,7 @@ import {
   type RemoteConnectionRequestMetadata,
 } from '../../remote/remote-connection-state.ts';
 import { profileToCliFlags } from '../remote-config-flags.ts';
-import type { BatchStep } from '../../client/client-types.ts';
+import type { BatchStep } from '../../contracts/client-api.ts';
 import { AppError } from '../../kernel/errors.ts';
 import type { LeaseBackend, SessionRuntimeHints } from '../../kernel/contracts.ts';
 import type { CliFlags } from '../../contracts/cli-flags.ts';
