+++ src/client/client-types.ts
@@ -1,72 +1,140 @@
-import type { PublicSnapshotCaptureAnnotations } from '../contracts/snapshot-capture-annotations.ts';
-import type { SnapshotDiagnosticsSummary } from '../contracts/snapshot-diagnostics.ts';
+// The published Node client surface.
+//
+// The VOCABULARY it is stated in terms of — connection config, the device/session views, and every
+// per-command Options/Result shape — lives in contracts/client-api.ts, below both this zone and
+// `commands/`. R2 forbids the reverse import, so a shape both surfaces need has to sit below both.
+//
+// What is still declared HERE is the `AgentDeviceClient` facade plus the shapes that are themselves
+// stated in terms of a HIGHER-ranked zone — `commands/` (navigation projection, ScrollInputDirection)
+// and `metro/` (prepare/reload results). Declaring those in contracts/ would trade 28
+// commands->client inversions for contracts->commands and contracts->metro ones: the foundation
+// depending on the layers above it, which is worse. They can move once their upstream declarations
+// do — see docs/dependency-graph-findings.md.
+//
+// The published surface is assembled in agent-device-client.ts, which re-exports BOTH homes, so no
+// name became unreachable from the package entrypoint.
+
+// The relocated vocabulary keeps its published import path through this module: one wildcard
+// rather than 84 named re-exports, so no name is published twice under two spellings.
+export type * from '../contracts/client-api.ts';
+
+// Contracts/kernel types re-exported into the PUBLISHED surface: `agent-device-client.ts` picks
+// these up via `export type *`, and that is their only job — every internal consumer imports
+// them from the declaring module instead. Fallow therefore sees no consumer, which is exactly
+// right and exactly not actionable: deleting them would remove names from the package's public
+// types. Suppressed per name rather than baselined so the reason travels with the code.
+// fallow-ignore-next-line unused-type
+export type { TargetShutdownResult } from '../contracts/target-shutdown-contract.ts';
+// fallow-ignore-next-line unused-type
+export type { MetroBridgeScope } from './client-companion-tunnel-contract.ts';
+// fallow-ignore-next-line unused-type
+export type { AppsFilter } from '../contracts/app-inventory.ts';
+// fallow-ignore-next-line unused-type
+export type { AlertAction } from '../contracts/alert-contract.ts';
+// fallow-ignore-next-line unused-type
+export type { AppleOS } from '../kernel/device.ts';
+// fallow-ignore-next-line unused-type
+export type { JsonObject } from '../contracts/json.ts';
+
 import type {
-  DaemonResponseData,
-  DaemonInstallSource,
-  DaemonLockPolicy,
-  DaemonRequest,
-  DaemonResponse,
-  LeaseBackend,
-  NetworkIncludeMode,
-  ResponseLevel,
-  SessionIsolationMode,
-  SessionRuntimeHints,
-} from '../kernel/contracts.ts';
-import type {
-  AppleOS,
-  DeviceKind,
-  DeviceTarget,
-  PublicPlatform,
-  PlatformSelector,
-} from '../kernel/device.ts';
-import type { BackMode } from '../contracts/back-mode.ts';
+  AgentDeviceCapabilitiesResult,
+  AgentDeviceClientConfig,
+  AgentDeviceDevice,
+  AgentDeviceRequestOverrides,
+  AgentDeviceSelectionOptions,
+  AgentDeviceSession,
+  AlertCommandOptions,
+  AppCloseOptions,
+  AppCloseResult,
+  AppDeployOptions,
+  AppDeployResult,
+  AppInstallFromSourceOptions,
+  AppInstallFromSourceResult,
+  AppInstallOptions,
+  AppListOptions,
+  AppOpenOptions,
+  AppOpenResult,
+  AppPushOptions,
+  AppStateCommandOptions,
+  AppTriggerEventOptions,
+  AudioOptions,
+  BatchRunOptions,
+  CaptureDiffOptions,
+  CaptureScreenshotOptions,
+  CaptureScreenshotResult,
+  CaptureSnapshotOptions,
+  CaptureSnapshotResult,
+  ClickOptions,
+  ClipboardCommandOptions,
+  CloudArtifactsOptions,
+  CommandRequestResult,
+  DeviceBootOptions,
+  DeviceCommandBaseOptions,
+  DeviceShutdownOptions,
+  DoctorCommandOptions,
+  EventsOptions,
+  FillOptions,
+  FindOptions,
+  FlingOptions,
+  FocusOptions,
+  GetOptions,
+  IsOptions,
+  KeyboardCommandOptions,
+  Lease,
+  LeaseAllocateOptions,
+  LeaseScopedOptions,
+  LogsOptions,
+  LongPressOptions,
+  MaterializationReleaseOptions,
+  MaterializationReleaseResult,
+  MetroPrepareOptions,
+  MetroReloadOptions,
+  NetworkOptions,
+  PanOptions,
+  PerfOptions,
+  PinchOptions,
+  PrepareCommandOptions,
+  PressOptions,
+  ReactNativeCommandOptions,
+  RecordOptions,
+  ReplayRunOptions,
+  ReplayTestOptions,
+  RotateGestureOptions,
+  SessionCloseResult,
+  SessionSaveScriptOptions,
+  SessionSaveScriptResult,
+  SettingsUpdateOptions,
+  SwipeGestureOptions,
+  SwipeOptions,
+  TraceOptions,
+  TransformGestureOptions,
+  TypeTextOptions,
+  ViewportCommandOptions,
+  WaitCommandOptions,
+} from '../contracts/client-api.ts';
+
 import type { RotateCommandResult } from '../contracts/navigation.ts';
-import type { ClickButton } from '../contracts/click-button.ts';
-import type { RecordingExportQuality } from '../contracts/recording-export-quality.ts';
-import type { RecordingScope } from '../contracts/recording-scope.ts';
-import type {
-  ScrollDirection,
-  SwipePattern,
-  SwipePreset,
-  TransformGestureParams,
-} from '../contracts/scroll-gesture.ts';
+
 import type { ScrollInputDirection } from '../commands/interaction/runtime/gestures.ts';
 import type {
   NavigationCommandOptions,
   ProjectedNavigationCommandClient,
 } from '../commands/system/navigation-projection.ts';
-import type { GesturePointerCount } from '../contracts/gesture-plan.ts';
-import type { LogAction } from '../contracts/logs.ts';
-import type { SessionSurface } from '../contracts/session-surface.ts';
-import type { FindLocator } from '../selectors/find.ts';
-import type { SnapshotNode, SnapshotUnchanged, SnapshotVisibility } from '../kernel/snapshot.ts';
-import type { ScreenshotResultData } from '../utils/screenshot-result.ts';
+
 import type { PrepareMetroRuntimeResult, ReloadMetroResult } from '../metro/client-metro.ts';
-import type { MetroPrepareKind } from '../contracts/metro.ts';
-import type { MetroBridgeScope } from './client-companion-tunnel-contract.ts';
-import type { AppsFilter } from '../contracts/app-inventory.ts';
-import type { ScreenshotRequestFlags } from '../contracts/screenshot.ts';
-import type { BatchRunResult, DaemonBatchStep } from '../core/batch.ts';
+
+import type { BatchRunResult } from '../core/batch.ts';
 export type { BatchRunResult } from '../core/batch.ts';
-import type { TargetShutdownResult } from '../contracts/target-shutdown-contract.ts';
-export type { TargetShutdownResult } from '../contracts/target-shutdown-contract.ts';
-import type { PerfAction, PerfArea, PerfKind, PerfSubject } from '../contracts/perf.ts';
-import type { AlertAction } from '../contracts/alert-contract.ts';
+
 import type { DebugSymbolsOptions, DebugSymbolsResult } from '../contracts/debug-symbols.ts';
-import type { JsonObject } from '../contracts/json.ts';
-import type { RemoteConnectionProfileFields } from '../remote/remote-config-schema.ts';
-import type { CloudProviderProfileFields } from '../contracts/remote-config-fields.ts';
+
 import type { CommandResult } from '../core/command-descriptor/command-result.ts';
 import type {
   AgentArtifactsResult,
   CloudProviderSessionResult,
 } from '../contracts/cloud-artifacts.ts';
 
-export type { MetroBridgeScope } from './client-companion-tunnel-contract.ts';
-export type { AppsFilter } from '../contracts/app-inventory.ts';
-export type { AlertAction } from '../contracts/alert-contract.ts';
 export type { DebugSymbolsOptions, DebugSymbolsResult } from '../contracts/debug-symbols.ts';
-export type { AppleOS } from '../kernel/device.ts';
 export type {
   /** @deprecated Renamed to `OrientationCommandResult`. Retained until the next major. */
   RotateCommandResult,
@@ -79,486 +147,10 @@ export type { DoctorCommandResult } from '../contracts/doctor.ts';
 export type { DiffSnapshotCommandResult } from '../contracts/diff.ts';
 export type { RecordingCommandResult, TraceCommandResult } from '../contracts/recording.ts';
 export type { ReplayCommandResult, ReplaySuiteResult } from '../contracts/replay.ts';
-export type { JsonObject } from '../contracts/json.ts';
-
-export type AgentDeviceDaemonTransport = (
-  req: Omit<DaemonRequest, 'token'>,
-) => Promise<DaemonResponse>;
-
-export type AgentDeviceClientConfig = RemoteConnectionProfileFields &
-  CloudProviderProfileFields & {
-    session?: string;
-    lockPolicy?: DaemonLockPolicy;
-    lockPlatform?: PlatformSelector;
-    requestId?: string;
-    sessionIsolation?: SessionIsolationMode;
-    leaseBackend?: LeaseBackend;
-    leaseTtlMs?: number;
-    runtime?: SessionRuntimeHints;
-    cwd?: string;
-    debug?: boolean;
-    cost?: boolean;
-    responseLevel?: ResponseLevel;
-    iosXctestrunFile?: string;
-    iosXctestDerivedDataPath?: string;
-    iosXctestEnvDir?: string;
-  };
-
-export type AgentDeviceRequestOverrides = Pick<
-  AgentDeviceClientConfig,
-  | 'session'
-  | 'lockPolicy'
-  | 'lockPlatform'
-  | 'requestId'
-  | 'daemonBaseUrl'
-  | 'daemonAuthToken'
-  | 'daemonTransport'
-  | 'daemonServerMode'
-  | 'tenant'
-  | 'sessionIsolation'
-  | 'runId'
-  | 'leaseId'
-  | 'leaseBackend'
-  | 'leaseProvider'
-  | 'deviceKey'
-  | 'clientId'
-  | 'providerApp'
-  | 'providerOsVersion'
-  | 'providerProject'
-  | 'providerBuild'
-  | 'providerSessionName'
-  | 'awsProjectArn'
-  | 'awsDeviceArn'
-  | 'awsAppArn'
-  | 'awsRegion'
-  | 'awsInteractionMode'
-  | 'leaseTtlMs'
-  | 'cwd'
-  | 'debug'
-  | 'cost'
-  | 'responseLevel'
-  | 'iosXctestrunFile'
-  | 'iosXctestDerivedDataPath'
-  | 'iosXctestEnvDir'
->;
-
-export type AgentDeviceIdentifiers = {
-  session?: string;
-  deviceId?: string;
-  deviceName?: string;
-  udid?: string;
-  serial?: string;
-  appId?: string;
-  appBundleId?: string;
-  package?: string;
-};
-
-export type AgentDeviceSelectionOptions = {
-  platform?: PlatformSelector;
-  target?: DeviceTarget;
-  device?: string;
-  udid?: string;
-  serial?: string;
-  iosSimulatorDeviceSet?: string;
-  androidDeviceAllowlist?: string;
-};
-
-export type AgentDeviceDevice = {
-  platform: PublicPlatform;
-  target: DeviceTarget;
-  kind: DeviceKind;
-  id: string;
-  name: string;
-  booted?: boolean;
-  /**
-   * Additive Apple-OS discriminant (iPhone/iPad/tvOS/visionOS/macOS). Present only for
-   * Apple devices; `platform` still carries the leaf (`ios`/`macos`).
-   */
-  appleOs?: AppleOS;
-  identifiers: AgentDeviceIdentifiers;
-  ios?: {
-    udid: string;
-  };
-  android?: {
-    serial: string;
-  };
-  vega?: {
-    serial: string;
-  };
-};
-
-export type AgentDeviceCapabilitiesResult = {
-  device: AgentDeviceDevice;
-  availableCommands: string[];
-};
-
-export type AgentDeviceSessionDevice = {
-  platform: PublicPlatform;
-  target: DeviceTarget;
-  id: string;
-  name: string;
-  /**
-   * Additive Apple-OS discriminant (iPhone/iPad/tvOS/visionOS/macOS). Present only for
-   * Apple devices; `platform` still carries the leaf (`ios`/`macos`).
-   */
-  appleOs?: AppleOS;
-  identifiers: AgentDeviceIdentifiers;
-  ios?: {
-    udid: string;
-    simulatorSetPath?: string | null;
-  };
-  android?: {
-    serial: string;
-  };
-  vega?: {
-    serial: string;
-  };
-};
-
-export type AgentDeviceSession = {
-  name: string;
-  createdAt: number;
-  sessionStateDir?: string;
-  runnerLogPath?: string;
-  device: AgentDeviceSessionDevice;
-  identifiers: AgentDeviceIdentifiers;
-};
-
-export type StartupPerfSample = {
-  durationMs: number;
-  measuredAt: string;
-  method: string;
-  appTarget?: string;
-  appBundleId?: string;
-};
-
-export type SessionCloseResult = {
-  session: string;
-  shutdown?: TargetShutdownResult;
-  provider?: CloudProviderSessionResult;
-  /**
-   * #1258: absolute path of the committed session/healed script when this close
-   * published one (`close --save-script`, or a repair-armed session's finalize)
-   * — so a client that requested publication learns where the file landed.
-   */
-  savedScript?: string;
-  identifiers: AgentDeviceIdentifiers;
-};
-
-export type SessionSaveScriptOptions = AgentDeviceRequestOverrides & {
-  path?: string;
-  /** Atomically replace an existing target instead of refusing publication. */
-  force?: boolean;
-};
-
-export type SessionSaveScriptResult = {
-  session: string;
-  savedScript: string;
-  actionCount: number;
-  identifiers: AgentDeviceIdentifiers;
-};
-
-export type CloudArtifactsOptions = AgentDeviceRequestOverrides & {
-  provider?: string;
-  providerSessionId?: string;
-};
-
-export type AppInstallOptions = AgentDeviceRequestOverrides &
-  AgentDeviceSelectionOptions & {
-    app?: string;
-    appPath: string;
-  };
-
-export type AppDeployOptions = AgentDeviceRequestOverrides &
-  AgentDeviceSelectionOptions & {
-    app: string;
-    appPath: string;
-  };
-
-export type AppDeployResult = {
-  app: string;
-  appPath: string;
-  platform: PublicPlatform;
-  appId?: string;
-  bundleId?: string;
-  package?: string;
-  identifiers: AgentDeviceIdentifiers;
-};
-
-export type AppOpenOptions = AgentDeviceRequestOverrides &
-  AgentDeviceSelectionOptions & {
-    app?: string;
-    url?: string;
-    surface?: SessionSurface;
-    activity?: string;
-    launchConsole?: string;
-    launchArgs?: string[];
-    relaunch?: boolean;
-    saveScript?: boolean | string;
-    /** #1258: overwrite an existing --save-script target instead of refusing. Alias: --overwrite. */
-    force?: boolean;
-    deviceHub?: boolean;
-    testIme?: boolean;
-    noRecord?: boolean;
-    runtime?: SessionRuntimeHints;
-  };
-
-export type AppOpenResult = {
-  session: string;
-  warnings?: string[];
-  sessionStateDir?: string;
-  runnerLogPath?: string;
-  requestLogPath?: string;
-  eventLogPath?: string;
-  appName?: string;
-  appBundleId?: string;
-  appId?: string;
-  startup?: StartupPerfSample;
-  runtime?: SessionRuntimeHints;
-  device?: AgentDeviceSessionDevice;
-  identifiers: AgentDeviceIdentifiers;
-};
-
-export type AppCloseOptions = AgentDeviceRequestOverrides & {
-  app?: string;
-  shutdown?: boolean;
-  saveScript?: boolean | string;
-  /** #1258: overwrite an existing --save-script target instead of refusing. Alias: --overwrite. */
-  force?: boolean;
-};
-
-export type AppCloseResult = {
-  session: string;
-  closedApp?: string;
-  shutdown?: TargetShutdownResult;
-  /**
-   * #1258: absolute path of the committed session/healed script when this close
-   * published one (`close --save-script`, or a repair-armed session's finalize)
-   * — so a client that requested publication learns where the file landed.
-   */
-  savedScript?: string;
-  identifiers: AgentDeviceIdentifiers;
-};
-
-export type AppInstallFromSourceOptions = AgentDeviceRequestOverrides &
-  AgentDeviceSelectionOptions & {
-    source: DaemonInstallSource;
-    retainPaths?: boolean;
-    retentionMs?: number;
-  };
-
-export type AppInstallFromSourceResult = {
-  appName?: string;
-  appId?: string;
-  bundleId?: string;
-  packageName?: string;
-  launchTarget: string;
-  installablePath?: string;
-  archivePath?: string;
-  materializationId?: string;
-  materializationExpiresAt?: string;
-  identifiers: AgentDeviceIdentifiers;
-};
-
-export type AppListOptions = AgentDeviceRequestOverrides &
-  AgentDeviceSelectionOptions & {
-    appsFilter?: AppsFilter;
-  };
-
-export type MaterializationReleaseOptions = AgentDeviceRequestOverrides & {
-  materializationId: string;
-};
-
-export type MaterializationReleaseResult = {
-  released: boolean;
-  materializationId: string;
-  identifiers: AgentDeviceIdentifiers;
-};
-
-export type Lease = {
-  leaseId: string;
-  tenantId: string;
-  runId: string;
-  backend: LeaseBackend;
-  leaseProvider?: string;
-  deviceKey?: string;
-  clientId?: string;
-  createdAt?: number;
-  heartbeatAt?: number;
-  expiresAt?: number;
-};
-
-export type LeaseOptions = AgentDeviceRequestOverrides &
-  AgentDeviceSelectionOptions & {
-    ttlMs?: number;
-  };
-
-export type LeaseAllocateOptions = LeaseOptions & {
-  tenant: string;
-  runId: string;
-  leaseBackend?: LeaseBackend;
-  leaseProvider?: string;
-  provider?: string;
-  deviceKey?: string;
-  clientId?: string;
-};
-
-export type LeaseScopedOptions = LeaseOptions & {
-  tenant?: string;
-  runId?: string;
-  leaseId: string;
-  leaseBackend?: LeaseBackend;
-  leaseProvider?: string;
-  provider?: string;
-  deviceKey?: string;
-  clientId?: string;
-};
-
-export type MetroPrepareOptions = {
-  projectRoot?: string;
-  kind?: MetroPrepareKind;
-  publicBaseUrl?: string;
-  proxyBaseUrl?: string;
-  bearerToken?: string;
-  bridgeScope?: MetroBridgeScope;
-  launchUrl?: string;
-  companionProfileKey?: string;
-  companionConsumerKey?: string;
-  port?: number;
-  listenHost?: string;
-  statusHost?: string;
-  startupTimeoutMs?: number;
-  probeTimeoutMs?: number;
-  reuseExisting?: boolean;
-  installDependenciesIfNeeded?: boolean;
-  runtimeFilePath?: string;
-  logPath?: string;
-};
 
 export type MetroPrepareResult = PrepareMetroRuntimeResult;
 
-export type MetroReloadOptions = {
-  metroHost?: string;
-  metroPort?: number;
-  bundleUrl?: string;
-  timeoutMs?: number;
-};
-
 export type MetroReloadResult = ReloadMetroResult;
-
-export type CaptureSnapshotOptions = AgentDeviceRequestOverrides &
-  AgentDeviceSelectionOptions & {
-    interactiveOnly?: boolean;
-    depth?: number;
-    scope?: string;
-    raw?: boolean;
-    forceFull?: boolean;
-    timeoutMs?: number;
-    /**
-     * #1271 stage 2 (ADR 0012 amendment): `snapshot` is observation-only and
-     * excluded from a repair-armed heal by default; `record` forces it
-     * through. Mutually exclusive with `noRecord`.
-     */
-    noRecord?: boolean;
-    record?: boolean;
-  };
-
-export type CaptureSnapshotResult = {
-  nodes: SnapshotNode[];
-  truncated: boolean;
-  appName?: string;
-  appBundleId?: string;
-  visibility?: SnapshotVisibility;
-  unchanged?: SnapshotUnchanged;
-  snapshotDiagnostics?: SnapshotDiagnosticsSummary;
-  identifiers: AgentDeviceIdentifiers;
-  /**
-   * ADR 0014: the response-level ref-frame epoch the plain node refs were minted
-   * from. A ref-issuing snapshot carries it ONCE (nodes stay plain `@e12` for the
-   * token budget); pair a ref with it (`@e12~s<refsGeneration>`) before a mutation.
-   */
-  refsGeneration?: number;
-  /**
-   * Digest response view only: a capped list of `{ ref, label? }` pairs taken
-   * from the full `nodes` tree so the MCP layer can still pin refs when the
-   * default-level `nodes` payload is intentionally omitted.
-   */
-  refs?: Array<{ ref: string; label?: string }>;
-} & PublicSnapshotCaptureAnnotations;
-
-export type CaptureScreenshotOptions = AgentDeviceRequestOverrides & {
-  path?: string;
-  overlayRefs?: boolean;
-  pixelDensity?: number;
-  fullscreen?: boolean;
-  maxSize?: number;
-  stabilize?: boolean;
-  normalizeStatusBar?: boolean;
-  surface?: SessionSurface;
-};
-
-export type CaptureScreenshotResult = ScreenshotResultData & {
-  path: string;
-  identifiers: AgentDeviceIdentifiers;
-};
-
-export type DeviceCommandBaseOptions = AgentDeviceRequestOverrides & AgentDeviceSelectionOptions;
-
-type WaitCommandTarget =
-  | {
-      durationMs: number;
-      text?: never;
-      ref?: never;
-      selector?: never;
-      stable?: never;
-      quietMs?: never;
-      timeoutMs?: never;
-    }
-  | (SelectorSnapshotCommandOptions & {
-      text: string;
-      durationMs?: never;
-      ref?: never;
-      selector?: never;
-      stable?: never;
-      quietMs?: never;
-      timeoutMs?: number;
-    })
-  | (SelectorSnapshotCommandOptions & {
-      ref: string;
-      durationMs?: never;
-      text?: never;
-      selector?: never;
-      stable?: never;
-      quietMs?: never;
-      timeoutMs?: number;
-    })
-  | (SelectorSnapshotCommandOptions & {
-      selector: string;
-      durationMs?: never;
-      text?: never;
-      ref?: never;
-      stable?: never;
-      quietMs?: never;
-      timeoutMs?: number;
-    })
-  | (SelectorSnapshotCommandOptions & {
-      stable: true;
-      durationMs?: never;
-      text?: never;
-      ref?: never;
-      selector?: never;
-      quietMs?: number;
-      timeoutMs?: number;
-    });
-
-export type WaitCommandOptions = DeviceCommandBaseOptions & WaitCommandTarget;
-
-export type AlertCommandOptions = DeviceCommandBaseOptions & {
-  action?: AlertAction;
-  timeoutMs?: number;
-};
-
-export type AppStateCommandOptions = DeviceCommandBaseOptions;
 
 export type BackCommandOptions = DeviceCommandBaseOptions & NavigationCommandOptions<'back'>;
 
@@ -571,40 +163,8 @@ export type RotateCommandOptions = OrientationCommandOptions;
 export type AppSwitcherCommandOptions = DeviceCommandBaseOptions &
   NavigationCommandOptions<'app-switcher'>;
 
-export type KeyboardCommandOptions = DeviceCommandBaseOptions & {
-  action?: 'status' | 'dismiss' | 'enter' | 'return';
-};
-
-export type ClipboardCommandOptions =
-  | (DeviceCommandBaseOptions & {
-      action: 'read';
-    })
-  | (DeviceCommandBaseOptions & {
-      action: 'write';
-      text: string;
-    });
-
 export type TvRemoteCommandOptions = DeviceCommandBaseOptions &
   NavigationCommandOptions<'tv-remote'>;
-
-export type ReactNativeCommandOptions = DeviceCommandBaseOptions & {
-  action: 'dismiss-overlay';
-};
-
-export type PrepareCommandOptions = DeviceCommandBaseOptions & {
-  action: 'ios-runner';
-  timeoutMs?: number;
-};
-
-export type DoctorCommandOptions = DeviceCommandBaseOptions & {
-  targetApp?: string;
-  remote?: boolean;
-};
-
-export type ViewportCommandOptions = DeviceCommandBaseOptions & {
-  width: number;
-  height: number;
-};
 
 type NonNavigationCommandClient = {
   wait: (options: WaitCommandOptions) => Promise<CommandResult<'wait'>>;
@@ -635,67 +195,6 @@ type DeprecatedCommandClient = {
   rotate: (options: RotateCommandOptions) => Promise<RotateCommandResult>;
 };
 
-type SelectorSnapshotCommandOptions = Pick<CaptureSnapshotOptions, 'depth' | 'scope' | 'raw'>;
-type FindSnapshotCommandOptions = Pick<CaptureSnapshotOptions, 'depth' | 'raw'>;
-
-type PointTarget = {
-  x: number;
-  y: number;
-  ref?: never;
-  selector?: never;
-  label?: never;
-};
-
-type RefTarget = {
-  ref: string;
-  label?: string;
-  x?: never;
-  y?: never;
-  selector?: never;
-};
-
-type SelectorTarget = {
-  selector: string;
-  x?: never;
-  y?: never;
-  ref?: never;
-  label?: never;
-};
-
-export type InteractionTarget = PointTarget | RefTarget | SelectorTarget;
-
-export type ElementTarget = RefTarget | SelectorTarget;
-
-type RepeatedPressOptions = {
-  count?: number;
-  intervalMs?: number;
-  holdMs?: number;
-  jitterPx?: number;
-  doubleTap?: boolean;
-};
-
-export type DeviceBootOptions = DeviceCommandBaseOptions & {
-  headless?: boolean;
-};
-
-export type DeviceShutdownOptions = DeviceCommandBaseOptions;
-
-export type AppPushOptions = DeviceCommandBaseOptions & {
-  app: string;
-  payload: string | JsonObject;
-};
-
-export type AppTriggerEventOptions = DeviceCommandBaseOptions & {
-  event: string;
-  payload?: JsonObject;
-};
-
-export type CaptureDiffOptions = DeviceCommandBaseOptions &
-  Pick<CaptureSnapshotOptions, 'interactiveOnly' | 'depth' | 'scope' | 'raw'> & {
-    kind: 'snapshot';
-    out?: string;
-  };
-
 /**
  * Opt-in (#1101): after the action, wait for the UI to go quiet and return the
  * settled diff vs the pre-action tree (`settle` on the result) in the same
@@ -704,89 +203,6 @@ export type CaptureDiffOptions = DeviceCommandBaseOptions &
  * 10s) when `settle` is true. A bare `timeoutMs` without `settle` is ignored
  * for compatibility; `settleQuietMs` still requires `settle`.
  */
-type SettleCommandOptions = {
-  settle?: boolean;
-  settleQuietMs?: number;
-  timeoutMs?: number;
-};
-
-export type ClickOptions = DeviceCommandBaseOptions &
-  SelectorSnapshotCommandOptions &
-  InteractionTarget &
-  RepeatedPressOptions &
-  SettleCommandOptions & {
-    button?: ClickButton;
-    /**
-     * Opt-in (#1047): return cheap post-action evidence (AX digest, node counts,
-     * changedFromBefore) in the response instead of requiring a follow-up
-     * snapshot to confirm the action had an effect.
-     */
-    verify?: boolean;
-  };
-
-export type PressOptions = DeviceCommandBaseOptions &
-  SelectorSnapshotCommandOptions &
-  InteractionTarget &
-  RepeatedPressOptions &
-  SettleCommandOptions & {
-    verify?: boolean;
-  };
-
-export type LongPressOptions = DeviceCommandBaseOptions &
-  SelectorSnapshotCommandOptions &
-  InteractionTarget &
-  SettleCommandOptions & {
-    durationMs?: number;
-  };
-
-export type SwipeOptions = DeviceCommandBaseOptions & {
-  from: { x: number; y: number };
-  to: { x: number; y: number };
-  count?: number;
-  pauseMs?: number;
-  pattern?: SwipePattern;
-};
-
-export type PanOptions = DeviceCommandBaseOptions & {
-  x: number;
-  y: number;
-  dx: number;
-  dy: number;
-  pointerCount?: GesturePointerCount;
-  durationMs?: number;
-};
-
-export type FlingOptions = DeviceCommandBaseOptions & {
-  direction: ScrollDirection;
-  x: number;
-  y: number;
-  distance?: number;
-};
-
-export type SwipeGestureOptions = DeviceCommandBaseOptions & {
-  preset: SwipePreset;
-};
-
-export type FocusOptions = DeviceCommandBaseOptions & {
-  x: number;
-  y: number;
-};
-
-export type TypeTextOptions = DeviceCommandBaseOptions & {
-  text: string;
-  delayMs?: number;
-};
-
-export type FillOptions = DeviceCommandBaseOptions &
-  SelectorSnapshotCommandOptions &
-  InteractionTarget &
-  SettleCommandOptions & {
-    text: string;
-    delayMs?: number;
-    /** Publish this fill value as `${VAR}` when script recording is armed. */
-    recordAs?: string;
-    verify?: boolean;
-  };
 
 export type ScrollOptions = DeviceCommandBaseOptions & {
   direction: ScrollInputDirection;
@@ -795,335 +211,12 @@ export type ScrollOptions = DeviceCommandBaseOptions & {
   durationMs?: number;
 };
 
-export type PinchOptions = DeviceCommandBaseOptions & {
-  scale: number;
-  x?: number;
-  y?: number;
-};
-
-export type RotateGestureOptions = DeviceCommandBaseOptions & {
-  degrees: number;
-  x?: number;
-  y?: number;
-};
-
-export type TransformGestureOptions = DeviceCommandBaseOptions & TransformGestureParams;
-
 /**
  * #1271 stage 2 (ADR 0012 amendment): `get`/`is`/`find` are observation-only
  * and excluded from a repair-armed heal by default. `record` forces this
  * action through (the corrective-read case); `noRecord` continues to opt the
  * action out entirely. Mutually exclusive.
  */
-type RecordControlOptions = {
-  noRecord?: boolean;
-  record?: boolean;
-};
-
-export type GetOptions = DeviceCommandBaseOptions &
-  SelectorSnapshotCommandOptions &
-  ElementTarget &
-  RecordControlOptions & {
-    format: 'text' | 'attrs';
-  };
-
-type IsTextPredicateOptions = DeviceCommandBaseOptions &
-  SelectorSnapshotCommandOptions &
-  RecordControlOptions & {
-    predicate: 'text';
-    selector: string;
-    value: string;
-  };
-
-type IsStatePredicateOptions = DeviceCommandBaseOptions &
-  SelectorSnapshotCommandOptions &
-  RecordControlOptions & {
-    predicate: 'visible' | 'hidden' | 'exists' | 'editable' | 'selected' | 'focused';
-    selector: string;
-    value?: never;
-  };
-
-export type IsOptions = IsTextPredicateOptions | IsStatePredicateOptions;
-
-type FindBaseOptions = DeviceCommandBaseOptions &
-  FindSnapshotCommandOptions &
-  RecordControlOptions & {
-    locator?: FindLocator;
-    query: string;
-    first?: boolean;
-    last?: boolean;
-  };
-
-export type FindOptions =
-  | (FindBaseOptions & { action?: 'click' | 'focus' | 'exists' | 'getText' | 'getAttrs' })
-  | (FindBaseOptions & { action: 'wait'; timeoutMs?: number })
-  | (FindBaseOptions & { action: 'fill' | 'type'; value: string });
-
-export type ReplayRunOptions = AgentDeviceRequestOverrides &
-  AgentDeviceSelectionOptions & {
-    path: string;
-    runtime?: SessionRuntimeHints;
-    /**
-     * @deprecated ADR 0012 migration step 6: `--update` no longer rewrites
-     * the script. Accepted for backward compatibility; every divergence
-     * already carries ranked selector suggestions regardless of this flag.
-     */
-    update?: boolean;
-    /** @deprecated Use backend: 'maestro'. */
-    maestro?: boolean;
-    backend?: string;
-    env?: string[];
-    timeoutMs?: number;
-    /**
-     * ADR 0012 decision 4 / migration step 5: resume at this 1-based plan
-     * step, skipping `1..resumeFrom-1` without executing them. Requires
-     * `resumePlanDigest` from the divergence report that reported this
-     * step as the failure. `replay` only — `test` has no resume fields.
-     */
-    resumeFrom?: number;
-    /** The `resume.planDigest` from the divergence report `resumeFrom` came from. */
-    resumePlanDigest?: string;
-    /**
-     * ADR 0012 decision 6, R1/R6: arms agent-supervised re-record repair
-     * from this replay attempt onward. Optional string value is the healed
-     * `.ad`'s output path; absent one, it defaults to the `<path>` sibling
-     * `<stem>.healed.ad` when the repair ends with `close --save-script`.
-     */
-    saveScript?: boolean | string;
-    /** #1258: overwrite an existing --save-script target instead of refusing. Alias: --overwrite. */
-    force?: boolean;
-  };
-
-export type ReplayTestOptions = AgentDeviceRequestOverrides &
-  AgentDeviceSelectionOptions & {
-    paths: string[];
-    runtime?: SessionRuntimeHints;
-    update?: boolean;
-    /** @deprecated Use backend: 'maestro'. */
-    maestro?: boolean;
-    backend?: string;
-    env?: string[];
-    failFast?: boolean;
-    timeoutMs?: number;
-    retries?: number;
-    recordVideo?: boolean;
-    artifactsDir?: string;
-    /** @deprecated Use the CLI --reporter junit:<path> or --report-junit <path>. */
-    reportJunit?: string;
-    shardAll?: number;
-    shardSplit?: number;
-  };
-
-export type BatchStep = {
-  command: string;
-  input: Record<string, unknown>;
-  runtime?: SessionRuntimeHints;
-};
-
-export type BatchRunOptions = AgentDeviceRequestOverrides & {
-  steps: BatchStep[];
-  onError?: 'stop';
-  maxSteps?: number;
-  out?: string;
-};
-
-export type PerfOptions = DeviceCommandBaseOptions & {
-  area?: PerfArea;
-  subject?: PerfSubject;
-  action?: PerfAction;
-  kind?: PerfKind;
-  template?: string;
-  out?: string;
-  tracePath?: string;
-};
-
-export type LogsOptions = AgentDeviceRequestOverrides & {
-  action?: LogAction;
-  message?: string;
-  restart?: boolean;
-};
-
-export type EventsOptions = AgentDeviceRequestOverrides & {
-  cursor?: string;
-  limit?: number;
-};
-
-export type NetworkOptions = AgentDeviceRequestOverrides & {
-  action?: 'dump' | 'log';
-  limit?: number;
-  include?: NetworkIncludeMode;
-};
-
-export type AudioOptions = AgentDeviceRequestOverrides & {
-  action?: 'probe';
-  probeAction?: 'start' | 'status' | 'stop';
-  durationMs?: number;
-  bucketMs?: number;
-};
-
-export type RecordOptions = AgentDeviceRequestOverrides & {
-  action: 'start' | 'stop';
-  path?: string;
-  fps?: number;
-  maxSize?: number;
-  quality?: RecordingExportQuality;
-  hideTouches?: boolean;
-  recordingScope?: RecordingScope;
-};
-
-export type TraceOptions = AgentDeviceRequestOverrides & {
-  action: 'start' | 'stop';
-  path?: string;
-};
-
-export type PermissionTarget =
-  | 'camera'
-  | 'microphone'
-  | 'photos'
-  | 'contacts'
-  | 'contacts-limited'
-  | 'notifications'
-  | 'calendar'
-  | 'location'
-  | 'location-always'
-  | 'media-library'
-  | 'motion'
-  | 'reminders'
-  | 'siri'
-  | 'accessibility'
-  | 'screen-recording'
-  | 'input-monitoring';
-
-export type SettingsUpdateOptions =
-  | (DeviceCommandBaseOptions & {
-      setting: 'clear-app-state';
-      state: 'clear';
-      app?: string;
-    })
-  | (DeviceCommandBaseOptions & {
-      setting: 'wifi' | 'airplane' | 'location';
-      state: 'on' | 'off';
-    })
-  | (DeviceCommandBaseOptions & {
-      setting: 'location';
-      state: 'set';
-      latitude: number;
-      longitude: number;
-    })
-  | (DeviceCommandBaseOptions & {
-      setting: 'animations';
-      state: 'on' | 'off';
-    })
-  | (DeviceCommandBaseOptions & {
-      setting: 'appearance';
-      state: 'light' | 'dark' | 'toggle';
-    })
-  | (DeviceCommandBaseOptions & {
-      setting: 'faceid' | 'touchid';
-      state: 'match' | 'nonmatch' | 'enroll' | 'unenroll';
-    })
-  | (DeviceCommandBaseOptions & {
-      setting: 'fingerprint';
-      state: 'match' | 'nonmatch';
-    })
-  | (DeviceCommandBaseOptions & {
-      setting: 'permission';
-      state: 'grant' | 'deny' | 'reset';
-      permission: PermissionTarget;
-      mode?: 'full' | 'limited';
-    });
-
-type CommandExecutionOptions = Partial<ScreenshotRequestFlags> & {
-  positionals?: string[];
-  kind?: string;
-  out?: string;
-  artifact?: string;
-  dsym?: string;
-  searchPath?: string;
-  interactiveOnly?: boolean;
-  depth?: number;
-  scope?: string;
-  raw?: boolean;
-  forceFull?: boolean;
-  count?: number;
-  fps?: number;
-  maxSize?: number;
-  recordingScope?: RecordingScope;
-  quality?: RecordingExportQuality;
-  hideTouches?: boolean;
-  intervalMs?: number;
-  delayMs?: number;
-  durationMs?: number;
-  holdMs?: number;
-  jitterPx?: number;
-  pixels?: number;
-  doubleTap?: boolean;
-  verify?: boolean;
-  settle?: boolean;
-  settleQuietMs?: number;
-  clickButton?: ClickButton;
-  pauseMs?: number;
-  pattern?: SwipePattern;
-  headless?: boolean;
-  restart?: boolean;
-  replayUpdate?: boolean;
-  replayBackend?: string;
-  replayEnv?: string[];
-  replayShellEnv?: Record<string, string>;
-  replayFrom?: number;
-  replayPlanDigest?: string;
-  failFast?: boolean;
-  timeoutMs?: number;
-  retries?: number;
-  recordVideo?: boolean;
-  artifactsDir?: string;
-  shardAll?: number;
-  shardSplit?: number;
-  findFirst?: boolean;
-  findLast?: boolean;
-  networkInclude?: NetworkIncludeMode;
-  batchOnError?: 'stop';
-  batchMaxSteps?: number;
-  batchSteps?: DaemonBatchStep[];
-};
-
-export type InternalRequestOptions = AgentDeviceClientConfig &
-  AgentDeviceSelectionOptions &
-  CommandExecutionOptions & {
-    runtime?: SessionRuntimeHints;
-    overlayRefs?: boolean;
-    surface?: SessionSurface;
-    activity?: string;
-    launchConsole?: string;
-    launchArgs?: string[];
-    relaunch?: boolean;
-    shutdown?: boolean;
-    saveScript?: boolean | string;
-    /** #1258: overwrite an existing --save-script target instead of refusing. Alias: --overwrite. */
-    force?: boolean;
-    deviceHub?: boolean;
-    testIme?: boolean;
-    noRecord?: boolean;
-    /** Fill-only script parameter name used to publish `${VAR}` instead of literal text. */
-    recordAs?: string;
-    /** #1271 stage 2: force-record this action; mutually exclusive with `noRecord`. */
-    record?: boolean;
-    backMode?: BackMode;
-    metroHost?: string;
-    metroPort?: number;
-    bundleUrl?: string;
-    launchUrl?: string;
-    appsFilter?: AppsFilter;
-    installSource?: DaemonInstallSource;
-    retainMaterializedPaths?: boolean;
-    materializedPathRetentionMs?: number;
-    materializationId?: string;
-    leaseTtlMs?: number;
-    provider?: string;
-    providerSessionId?: string;
-  };
-
-export type CommandRequestResult = DaemonResponseData;
 
 export type AgentDeviceClient = {
   command: AgentDeviceCommandClient;
