+++ src/commands/capture/alert.ts
@@ -1,6 +1,6 @@
 import { ALERT_ACTIONS, type AlertAction } from '../../contracts/alert-contract.ts';
 import { PUBLIC_COMMANDS } from '../../command-catalog.ts';
-import type { AlertCommandOptions } from '../../client/client-types.ts';
+import type { AlertCommandOptions } from '../../contracts/client-api.ts';
 import { compactRecord, enumField, integerField } from '../command-input.ts';
 import { defineExecutableCommand } from '../command-contract.ts';
 import {
