+++ src/commands/command-flags.ts
@@ -3,7 +3,7 @@ import type { CommandFlags } from '../core/dispatch-context.ts';
 import { leaseScopeFromOptions, leaseScopeToCommandFlags } from '../core/lease-scope.ts';
 import { stripUndefined } from '../utils/parsing.ts';
 import { getFlagDefinitions } from './cli-grammar/flag-registry.ts';
-import type { InternalRequestOptions } from '../client/client-types.ts';
+import type { InternalRequestOptions } from '../contracts/client-api.ts';
 import type { CommandMetadata } from './command-contract.ts';
 
 const CLI_FLAG_KEYS: ReadonlySet<string> = new Set(
