+++ src/commands/interaction/selectors.ts
@@ -1,5 +1,5 @@
 import { PUBLIC_COMMANDS } from '../../command-catalog.ts';
-import type { FindOptions, IsOptions } from '../../client/client-types.ts';
+import type { FindOptions, IsOptions } from '../../contracts/client-api.ts';
 import type { CliFlags } from '../../contracts/cli-flags.ts';
 import { AppError } from '../../kernel/errors.ts';
 import { checkIsPredicate, normalizeIsPositionals } from '../../selectors/predicates.ts';
