+++ src/commands/management/app.ts
@@ -1,5 +1,5 @@
 import { PUBLIC_COMMANDS } from '../../command-catalog.ts';
-import type { AppCloseOptions, AppOpenOptions } from '../../client/client-types.ts';
+import type { AppCloseOptions, AppOpenOptions } from '../../contracts/client-api.ts';
 import { DEFAULT_APPS_FILTER } from '../../contracts/app-inventory.ts';
 import { SESSION_SURFACES } from '../../contracts/session-surface.ts';
 import type { CommandSchemaOverride } from '../../cli-schema/types.ts';
