+++ src/commands/metro/index.ts
@@ -1,9 +1,5 @@
-import type {
-  MetroPrepareOptions,
-  MetroPrepareResult,
-  MetroReloadOptions,
-  MetroReloadResult,
-} from '../../client/client-types.ts';
+import type { MetroPrepareOptions, MetroReloadOptions } from '../../contracts/client-api.ts';
+import type { MetroPrepareResult, MetroReloadResult } from '../../client/client-types.ts';
 import { AppError } from '../../kernel/errors.ts';
 import type { CommandSchemaOverride } from '../../cli-schema/types.ts';
 import {
