+++ src/commands/observability/index.ts
@@ -3,7 +3,7 @@ import type {
   EventsOptions,
   LogsOptions,
   NetworkOptions,
-} from '../../client/client-types.ts';
+} from '../../contracts/client-api.ts';
 import { NETWORK_INCLUDE_MODES, type NetworkIncludeMode } from '../../kernel/contracts.ts';
 import { AppError } from '../../kernel/errors.ts';
 import { parseStringMember } from '../../utils/string-enum.ts';
