+++ src/commands/system/index.ts
@@ -1,4 +1,4 @@
-import type { ClipboardCommandOptions } from '../../client/client-types.ts';
+import type { ClipboardCommandOptions } from '../../contracts/client-api.ts';
 import type { BackMode } from '../../contracts/back-mode.ts';
 import { BACK_MODES } from '../../contracts/back-mode.ts';
 import { parseDeviceRotation, DEVICE_ROTATIONS } from '../../contracts/device-rotation.ts';
