+++ src/mcp/__tests__/command-tools-parity.test.ts
@@ -4,7 +4,8 @@ import os from 'node:os';
 import path from 'node:path';
 import { afterEach, test, vi } from 'vitest';
 import { createAgentDeviceClient } from '../../agent-device-client.ts';
-import type { AgentDeviceClient, AgentDeviceDaemonTransport } from '../../client/client-types.ts';
+import type { AgentDeviceDaemonTransport } from '../../contracts/client-api.ts';
+import type { AgentDeviceClient } from '../../client/client-types.ts';
 import type { CommandExecutionResult } from '../../commands/command-surface.ts';
 import { createCommandToolExecutor, listCommandTools } from '../command-tools.ts';
 import { validateAgainstSchema } from './output-schema-validator.ts';
