+++ test/output-economy/fixtures.ts
@@ -1,4 +1,7 @@
-import type { CaptureSnapshotResult, CommandRequestResult } from '../../src/client/client-types.ts';
+import type {
+  CaptureSnapshotResult,
+  CommandRequestResult,
+} from '../../src/contracts/client-api.ts';
 import type { DaemonResponseData } from '../../src/daemon/types.ts';
 import { AppError } from '../../src/kernel/errors.ts';
 import { attachRefs, type RawSnapshotNode } from '../../src/kernel/snapshot.ts';
